+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Fix: Pinned to `textual<0.2.0` due to breaking changes 0.6.0.dev13 #107
 * Fix: snyk remove seuptools from requirements 0.6.0.dev14 #130
 * Fix: Markata.filter was missing post and m 0.6.0.dev14 #133
+* Feat: add `path_prefix` config for gh-pages deploy 0.6.0.dev15 #132 fixes #57
 
 ## 0.5.2